[Refactor] domain 컴포넌트 폴더 구조 컨벤션 적용 - components/domain/content → components/domain/tu/content 이동 - 역할별(sa/ta/to/tu) 폴더 구조 컨벤션 준수 - import 경로 수정
