+++ src/pages/tu/teaching/content/ContentCreatePage.tsx
@@ -1,5 +1,5 @@
 import { useNavigate } from 'react-router-dom';
-import { ContentRegistrationWizard } from '@/components/domain/content';
+import { ContentRegistrationWizard } from '@/components/domain/tu/content';
 import type { LOData } from '@/types';
 
 export function ContentCreatePage() {
